Split conduit command tests from conduit helper function tests
